requirements.txt
# AI Scheduling Assistant - Dependencies
# Install with: pip install -r requirements.txt

# Web Framework
flask>=3.0.0

# OpenAI
openai>=1.0.0

# Twilio
twilio>=9.0.0

# Google Calendar API
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.0

# Environment & Config
python-dotenv>=1.0.0

# Date/Time Parsing
python-dateutil>=2.8.0

# Timezone Data (required on Windows)
tzdata>=2024.1
